SlackApi/requirements.txt
azure-functions
slack